WPF: Use a DependencyProperty
Move the MediaPlayerEventManager in LibVLCSharp since it's used by WPF in addition to Forms
