+++ LibVLCSharp.Forms/Shared/VideoView.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using LibVLCSharp.Shared;
+using System;
 using System.Diagnostics;
 using Xamarin.Forms;
